fix the lint on application

+++ nature_gift_admin/app/api/blogs/[blogId]/route.ts
@@ -1,3 +1,4 @@
+import { getUserByClerkId } from '@/lib/actions/actions'
 import { blogSchema } from '@/lib/validations/blog'
 import { auth } from '@clerk/nextjs/server'
 import { prisma } from '@naturegift/models'
@@ -6,8 +7,18 @@ import { NextResponse } from 'next/server'
 export async function GET(req: Request, props: { params: Promise<{ blogId: string }> }) {
   const params = await props.params
   try {
+    const { userId } = await auth()
+
+    if (!userId) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+    }
+    const _currentUser = await getUserByClerkId(userId)
+    if (!_currentUser?.id) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+    }
+
     const blog = await prisma.blog.findUnique({
-      where: { id: params.blogId },
+      where: { id: params.blogId, creatorId: _currentUser.id },
       include: {
         categories: {
           include: {
@@ -42,6 +53,11 @@ export async function PATCH(req: Request, props: { params: Promise<{ blogId: str
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
     }
 
+    const _currentUser = await getUserByClerkId(userId)
+    if (!_currentUser?.id) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+    }
+
     const json = await req.json()
 
     const body = blogSchema.partial().parse(json)
@@ -50,7 +66,7 @@ export async function PATCH(req: Request, props: { params: Promise<{ blogId: str
     const { categoryIds: categoryIds, ...updateData } = body
 
     const blog = await prisma.blog.update({
-      where: { id: params.blogId },
+      where: { id: params.blogId, creatorId: _currentUser.id },
       data: {
         ...updateData,
         customFields: [],
@@ -93,8 +109,13 @@ export async function DELETE(req: Request, props: { params: Promise<{ blogId: st
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
     }
 
+    const _currentUser = await getUserByClerkId(userId)
+    if (!_currentUser?.id) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+    }
+
     await prisma.blog.delete({
-      where: { id: params.blogId },
+      where: { id: params.blogId, creatorId: _currentUser.id },
     })
 
     return NextResponse.json({ message: 'Blog deleted successfully' })
